pyproject.toml
[project]
name = "consilium"
version = "0.1.0"
description = "Multi-LLM council for developer product concept work"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.27",
    "pydantic>=2.7",
    "pyyaml>=6.0",
    "tiktoken>=0.7",
    "anthropic>=0.40",
    "pymupdf>=1.24",
    "python-docx>=1.1",
    "aiosqlite>=0.20",
]

[project.optional-dependencies]
server = [
    "aiogram>=3.4",
    "fastapi>=0.111",
    "uvicorn[standard]>=0.29",
]
api = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "sse-starlette>=2.1",
    "httpx>=0.27",  # for FastAPI TestClient
    "python-multipart>=0.0.9",  # for multipart/form-data (pack upload)
]
bot = [
    "aiogram>=3.13",
    "httpx>=0.27",
]
mcp = [
    "mcp>=1.0",
]
cli = [
    "rich>=13",
    "httpx>=0.27",
]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.23",
    "respx>=0.21",
    "ruff>=0.4",
    "mypy>=1.10",
    "types-PyYAML>=6.0",
]

[project.scripts]
consilium = "consilium_cli.main:main"
consilium-mcp = "consilium_mcp.server:main"
consilium-bot = "consilium_server.bot.main:main"
consilium-api = "consilium_server.api.main:main"
consilium-budget = "scripts.budget:main"

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = ["integration: real API calls, skipped by default"]
addopts = "-m 'not integration'"

[tool.ruff]
line-length = 100
target-version = "py312"

[[tool.mypy.overrides]]
module = ["fitz", "docx", "docx.*"]
ignore_missing_imports = true

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = [
    "consilium*",
    "consilium_server*",
    "consilium_mcp*",
    "consilium_cli*",
    "scripts*",
]
